Move recovery related functions to separate module
