+++ src/reparsec/core/combinators.py
@@ -1,118 +1,17 @@
 from typing import Callable, List, Optional, Tuple, TypeVar
 
-from .chain import Append, Cons
-from .result import Error, Ok, PrefixItem, Recovered, Repair, Result, Selected
+from .chain import Append
+from .recovery import MergeFn, continue_parse, join_repairs
+from .result import Error, Ok, Recovered, Repair, Result, Selected
 from .state import Ctx
 from .types import (
     ParseFn, ParseObj, RecoveryMode, disallow_recovery, maybe_allow_recovery
 )
 
 S = TypeVar("S")
-S_contra = TypeVar("S_contra", contravariant=True)
 V = TypeVar("V")
-V_co = TypeVar("V_co", covariant=True)
 U = TypeVar("U")
 X = TypeVar("X")
-
-
-ContinueFn = Callable[[V, S, int, Ctx[S]], Result[U, S]]
-MergeFn = Callable[[V, U], X]
-
-
-def continue_parse(
-        stream: S, ra: Recovered[V, S], parse: ContinueFn[V, S, U],
-        merge: MergeFn[V, U, X]) -> Result[X, S]:
-
-    def _handle_pending(rep: Repair[V, S]) -> Optional[Selected[X, S]]:
-        rb = parse(rep.value, stream, rep.pos, rep.ctx)
-        if type(rb) is Ok:
-            if selected is None or selected.selected > rep.pos:
-                return Selected(
-                    rep.pos, merge(rep.value, rb.value), rb.pos, rb.ctx,
-                    rep.op, rep.expected, rep.consumed or rb.consumed,
-                    rep.prefix
-                )
-        elif type(rb) is Recovered:
-            pending.extend(
-                Repair(
-                    merge(rep.value, pb.value), pb.pos, pb.ctx, pb.op,
-                    pb.expected, pb.consumed,
-                    Append(
-                        rep.prefix,
-                        Cons(PrefixItem(rep.op, rep.expected), pb.prefix)
-                    )
-                )
-                for pb in rb.pending
-            )
-            if rb.selected is not None:
-                repb = rb.selected
-                if selected is None or selected.selected > repb.selected:
-                    return Selected(
-                        repb.selected, merge(rep.value, repb.value), repb.pos,
-                        repb.ctx, repb.op, repb.expected, repb.consumed,
-                        Append(
-                            rep.prefix,
-                            Cons(PrefixItem(rep.op, rep.expected), repb.prefix)
-                        )
-                    )
-        return selected
-
-    sel = ra.selected
-    if sel is not None:
-        rb = parse(sel.value, stream, sel.pos, sel.ctx)
-        selected, pending = _merge_selected(sel, rb, merge)
-    else:
-        selected = None
-        pending = []
-
-    for pa in ra.pending:
-        selected = _handle_pending(pa)
-
-    if selected is not None or pending:
-        return Recovered(selected, pending, ra.pos, ra.loc, ra.expected)
-
-    return Error(ra.pos, ra.loc, ra.expected, True)
-
-
-def _merge_selected(
-        rep: Selected[V, S], rb: Result[U, S],
-        merge: MergeFn[V, U, X]) -> Tuple[
-            Optional[Selected[X, S]], List[Repair[X, S]]]:
-    if type(rb) is Ok:
-        return (
-            Selected(
-                rep.selected, merge(rep.value, rb.value), rb.pos, rb.ctx,
-                rep.op, rep.expected, rep.consumed or rb.consumed, rep.prefix
-            ),
-            []
-        )
-    elif type(rb) is Recovered:
-        pending = [
-            Repair(
-                merge(rep.value, pb.value), pb.pos, pb.ctx, pb.op,
-                pb.expected, pb.consumed,
-                Append(
-                    rep.prefix,
-                    Cons(PrefixItem(rep.op, rep.expected), pb.prefix)
-                )
-            )
-            for pb in rb.pending
-        ]
-        if rb.selected is not None:
-            repb = rb.selected
-            return (
-                Selected(
-                    rep.selected, merge(rep.value, repb.value), repb.pos,
-                    repb.ctx, repb.op, repb.expected, repb.consumed,
-                    Append(
-                        rep.prefix,
-                        Cons(PrefixItem(rep.op, rep.expected), repb.prefix)
-                    )
-                ),
-                pending
-            )
-        return None, pending
-    return None, []
 
 
 def fmap(parse_fn: ParseFn[S, V], fn: Callable[[V], U]) -> ParseFn[S, U]:
@@ -143,7 +42,7 @@ def alt(parse_fn: ParseFn[S, V], second_fn: ParseFn[S, V]) -> ParseFn[S, V]:
             rrb = second_fn(stream, pos, ctx, True)
             if type(rra) is Recovered:
                 if type(rrb) is Recovered:
-                    return _merge_alt_recovered(rra, rrb).expect(expected)
+                    return join_repairs(rra, rrb).expect(expected)
                 return rra.expect(expected)
             if type(rrb) is Recovered:
                 return rrb.expect(expected)
